Implementar verificación limite reservas para el servcio y ajustes en DTO cliente, empresario y empresa

+++ src/main/java/com/sompoble/cat/domain/Servicio.java
@@ -137,7 +137,7 @@ public class Servicio implements Serializable {
         this.precio = precio;
     }
     
-    public float getLimiteReservas() {
+    public int getLimiteReservas() {
         return limiteReservas;
     }
